requirements.txt
Django==2.1.11
pyodbc>=4.0.21
django-pyodbc-azure==2.1.0.0
pandas>=0.21.0
sqlalchemy==1.3.6
pyjwt==1.7.1
cryptography==2.8.0
requests==2.22.0
simplejson==3.17.0
graphene-django==2.2.0
django-cors-headers>=3.2.0
xlrd>=1.2.0
geopy==1.20.0
msal==1.2.0
azure-storage-blob==12.3.2
azure-cosmosdb-table==1.0.6
pytz==2020.1